src/main/java/org/ryuu/util/NullOrEmptyUtils.java
package org.ryuu.util;

import lombok.NoArgsConstructor;

import java.util.Collection;
import java.util.Map;

@NoArgsConstructor
public class NullOrEmptyUtils {
    public static <T> boolean isNullOrEmpty(String string) {
        return string == null || string.isEmpty();
    }

    public static <T> boolean isNullOrEmpty(Collection<T> collection) {
        return collection == null || collection.isEmpty();
    }

    public static <T> boolean isNullOrEmpty(T[] array) {
        return array == null || array.length == 0;
    }

    public static <K, V> boolean isNullOrEmpty(Map<K, V> map) {
        return map == null || map.isEmpty();
    }
}
